+++ src/ProfileManagement/account.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import './App.css';
+import './profileStyle.css';
 
 const Account = () => {
     return(
-        <div className="container">
+        <div className="container-acc">
             <div className="row gutters">
                     <div className="col-xl-3 col-lg-3 col-md-12 col-sm-12 col-12">
                         <div className="card h-100">
