+++ src/ProfileManagement/bookhistory.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import './App.css';
+import './profileStyle.css';
 
-const Location = () => {
+const Booked = () => {
     return (
         <div class="container-xl">
             <div class="table-responsive">
@@ -10,7 +10,7 @@ const Location = () => {
                     <div class="table-title">
                         <div class="row">
                             <div class="col-sm-5">
-                                <h2>Favorite Professionals</h2>
+                                <h2>Bookings</h2>
                             </div>
                             
                         </div>
@@ -19,9 +19,9 @@ const Location = () => {
                         <thead>
                             <tr>
                                 <th>#</th>
-                                <th>Name</th>						
-                                <th>Profession</th>
-                                
+                                <th>Name</th>	
+                                <th>Service</th>					
+                                <th>Date of Service</th>
                                 <th>Action</th>
                             </tr>
                         </thead>
@@ -29,7 +29,8 @@ const Location = () => {
                             <tr>
                                 <td>1</td>
                                 <td><a href="#">Michael Holz</a></td>
-                                <td>Electrician</td>                        
+                                <td>Electrician</td>
+                                <td>04/10/2013</td>                        
                                 
                                 <td>
                                    
@@ -39,9 +40,9 @@ const Location = () => {
                             <tr>
                                 <td>2</td>
                                 <td><a href="#"> Paula Wilson</a></td>
-                                <td>
-                                Home Cleaning
-                                </td>                       
+                                
+                                <td>Home Cleaning</td>
+                                <td>04/10/2013</td>                       
                                 
                                 <td>
                                     
@@ -51,7 +52,8 @@ const Location = () => {
                             <tr>
                                 <td>3</td>
                                 <td><a href="#"> Antonio Moreno</a></td>
-                                <td> Plumber</td>
+                                <td>Plumbing</td>
+                                <td>11/05/2015</td>
                                                        
                                 <td>
                                     
@@ -61,7 +63,8 @@ const Location = () => {
                             <tr>
                                 <td>4</td>
                                 <td><a href="#"> Mary Saveley</a></td>
-                                <td> Home Cleaning </td>
+                                <td>Home Cleaning</td>
+                                <td>06/09/2016</td>
                                 
                                 <td>
                                     
@@ -71,11 +74,12 @@ const Location = () => {
                             <tr>
                                 <td>5</td>
                                 <td><a href="#"> Martin Sommer</a></td>
-                                <td> Aircon Cleaner </td>                        
+                                <td>Aircon Cleaning</td>
+                                <td>12/08/2017</td>                        
                                 
                                 <td>
                                     
-                                    <button type="button" class="btn-close" aria-label="Close"></button>
+                                <button type="button" class="btn-close" aria-label="Close"></button>
                                 </td>
                             </tr>
                         </tbody>
@@ -97,5 +101,4 @@ const Location = () => {
         </div>   
     );
 }
-
-export default Location;
+export default Booked;
