+++ src/ProfileManagement/profileStyle.css
@@ -1,0 +1,379 @@
+.App {
+    height: 100vh;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #f5f6fa;
+    background-image: url(./images/bck.jpg);
+  }
+  
+  .cfb{
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  /* Container Styling */
+  .Container{
+    position: relative; 
+    margin-top: 60px;
+    width: 50%;
+    height: 75%;
+  
+    box-shadow: 0 0.188em 1.550em rgb(156, 156, 156);
+    
+    border-radius: 0.5em;
+  
+   
+  }
+  
+  /* Banner Side of Container */
+  .banner-side{
+    position: absolute;
+    z-index: 2;
+    right: 65%;
+  
+    border: 0.1em solid #592941;
+    border-radius: 0.5em;
+  
+    width: 35%;
+    height: 100%;
+  
+    flex-direction: column;
+  
+    background-color: blueviolet;
+    opacity: 1;
+  
+    color: black;
+    text-align: center;
+  
+    transition: 2s;
+  }
+  
+  /* Form Side of Container */
+  .form-side{
+    position: absolute;
+    z-index: 1;
+    right: 0;
+  
+    border: 0.1em solid #592941;
+    border-radius: 0.5em;
+  
+    width: 65%;
+    height: 100%;
+  
+    background-color: #EEFCF8;
+    opacity: 1;
+  
+    font-weight: bold;
+    color: black;
+  
+    transition: 2s;
+  }
+  
+  .send-right{
+    right: 0;
+  }
+  
+  .send-left{
+    right: 35%;
+  }
+  
+  
+  /* Layout styling of forms, labels, inputs */
+  .form-comp{
+    flex-direction: column;
+  }
+  
+  .sign-up-form{
+    width: 50%;
+    flex-direction: column;
+  }
+  
+  .sign-up-form > label{
+    padding: 0.5em;
+    width: 100%
+  }
+  
+  .sign-up-form > label > input{
+    width: 100%
+  }
+  
+  
+  /* All Buttons Styling */
+  button {
+    padding: 1em;
+  
+    border: 1px solid #EEFCF8;
+    border-radius: 5em;
+    
+    background-color: #035E7B;
+  
+    color: #EEFCF8;
+    font-weight: bold;
+    text-decoration: none;
+  }
+  
+  button:focus {
+    outline: 0;
+  }
+  
+  button:hover{
+    opacity: 0.9;
+  }
+  
+  /* All Inputs Styling */
+  input {
+    padding: 0.5em;
+    border: 1px solid #035E7B;
+  }
+  
+  /* Account Management */
+  .container-acc{
+    margin-top: 60px;
+  }
+  .account-settings .user-profile {
+    margin: 0 0 1rem 0;
+    padding-bottom: 1rem;
+    text-align: center;
+  }
+  .account-settings .user-profile .user-avatar {
+    margin: 0 0 1rem 0;
+  }
+  .account-settings .user-profile .user-avatar img {
+    width: 90px;
+    height: 90px;
+    -webkit-border-radius: 100px;
+    -moz-border-radius: 100px;
+    border-radius: 100px;
+  }
+  .account-settings .user-profile h5.user-name {
+    margin: 0 0 0.5rem 0;
+  }
+  .account-settings .user-profile h6.user-email {
+    margin: 0;
+    font-size: 0.8rem;
+    font-weight: 400;
+    color: #9fa8b9;
+  }
+  .account-settings .about {
+    margin: 2rem 0 0 0;
+    text-align: center;
+  }
+  .account-settings .about h5 {
+    margin: 0 0 15px 0;
+    color: #007ae1;
+  }
+  .account-settings .about p {
+    font-size: 0.825rem;
+  }
+  .form-control {
+    border: 1px solid #cfd1d8;
+    -webkit-border-radius: 2px;
+    -moz-border-radius: 2px;
+    border-radius: 2px;
+    font-size: .825rem;
+    background: #ffffff;
+    color: #2e323c;
+  }
+  
+  .card {
+    background: #ffffff;
+    -webkit-border-radius: 5px;
+    -moz-border-radius: 5px;
+    border-radius: 5px;
+    border: 0;
+    margin-bottom: 1rem;
+  }
+  
+  .btnm{
+    margin-top: 20px;
+    margin-right: 10px;
+  }
+  .tempimg{
+    background-image: url(./images/bean.jpg);
+    height: 200px;
+  }
+  
+  /* Professionals */
+  .container-xl{
+    margin-top: 75px;
+  }
+  .table-responsive {
+    margin: 30px 0;
+  }
+  .table-wrapper {
+    min-width: 1000px;
+    background: #fff;
+    padding: 20px 25px;
+    border-radius: 3px;
+    box-shadow: 0 1px 1px rgba(0,0,0,.05);
+  }
+  .table-title {
+    padding-bottom: 15px;
+    background: #299be4;
+    color: #fff;
+    padding: 16px 30px;
+    margin: -20px -25px 10px;
+    border-radius: 3px 3px 0 0;
+  }
+  .table-title h2 {
+    margin: 5px 0 0;
+    font-size: 24px;
+  }
+  .table-title .btn {
+    color: #566787;
+    float: right;
+    font-size: 13px;
+    background: #fff;
+    border: none;
+    min-width: 50px;
+    border-radius: 2px;
+    border: none;
+    outline: none !important;
+    margin-left: 10px;
+  }
+  .table-title .btn:hover, .table-title .btn:focus {
+    color: #566787;
+    background: #f2f2f2;
+  }
+  .table-title .btn i {
+    float: left;
+    font-size: 21px;
+    margin-right: 5px;
+  }
+  .table-title .btn span {
+    float: left;
+    margin-top: 2px;
+  }
+  table.table tr th, table.table tr td {
+    border-color: #e9e9e9;
+    padding: 12px 15px;
+    vertical-align: middle;
+  }
+  table.table tr th:first-child {
+    width: 60px;
+  }
+  table.table tr th:last-child {
+    width: 100px;
+  }
+  table.table-striped tbody tr:nth-of-type(odd) {
+    background-color: #fcfcfc;
+  }
+  table.table-striped.table-hover tbody tr:hover {
+    background: #f5f5f5;
+  }
+  table.table th i {
+    font-size: 13px;
+    margin: 0 5px;
+    cursor: pointer;
+  }	
+  table.table td:last-child i {
+    opacity: 0.9;
+    font-size: 22px;
+    margin: 0 5px;
+  }
+  table.table td a {
+    font-weight: bold;
+    color: #566787;
+    display: inline-block;
+    text-decoration: none;
+  }
+  table.table td a:hover {
+    color: #2196F3;
+  }
+  table.table td a.settings {
+    color: #2196F3;
+  }
+  table.table td a.delete {
+    color: #F44336;
+  }
+  table.table td i {
+    font-size: 19px;
+  }
+  table.table .avatar {
+    border-radius: 50%;
+    vertical-align: middle;
+    margin-right: 10px;
+  }
+  .status {
+    font-size: 30px;
+    margin: 2px 2px 0 0;
+    display: inline-block;
+    vertical-align: middle;
+    line-height: 10px;
+  }
+  .text-success {
+    color: #10c469;
+  }
+  .text-info {
+    color: #62c9e8;
+  }
+  .text-warning {
+    color: #FFC107;
+  }
+  .text-danger {
+    color: #ff5b5b;
+  }
+  .pagination {
+    float: right;
+    margin: 0 0 5px;
+  }
+  .pagination li a {
+    border: none;
+    font-size: 13px;
+    min-width: 30px;
+    min-height: 30px;
+    color: #999;
+    margin: 0 2px;
+    line-height: 30px;
+    border-radius: 2px !important;
+    text-align: center;
+    padding: 0 6px;
+  }
+  .pagination li a:hover {
+    color: #666;
+  }	
+  .pagination li.active a, .pagination li.active a.page-link {
+    background: #03A9F4;
+  }
+  .pagination li.active a:hover {        
+    background: #0397d6;
+  }
+  .pagination li.disabled i {
+    color: #ccc;
+  }
+  .pagination li i {
+    font-size: 16px;
+    padding-top: 6px
+  }
+  .hint-text {
+    float: left;
+    margin-top: 10px;
+    font-size: 13px;
+  }
+  .ratings{
+    margin-right:10px;
+  }
+  
+  .ratings i{
+    
+    color:#cecece;
+    font-size:32px;
+  }
+  
+  .rating-color{
+    color:#fbc634 !important;
+  }
+  
+  .small-ratings i{
+    color:#cecece;   
+  }
+  
+  
+  
+  
+  
+
