--- src/register.css
@@ -1,212 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap");
-@import url("https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css");
-
- *{
-    margin: 0px;
-    padding: 0px;
-    box-sizing: border-box;
-  }
-
-  .container {
-    background-color: #fff;
-    width: 760px;
-    max-width: 100vw;
-    height: 480px;
-    position: relative;
-    overflow-x: hidden;
-    box-shadow: 0 0.188em 1.550em rgb(156, 156, 156);
-  }
- 
-  .container .form-card{
-    position: relative;
-    width: 50%;
-    text-align: center;
-  }
- 
-  .container .form-card .form-control {
-    position: absolute;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    height: 480px;
-    transition: all 0.5s ease-in;
-  }
-
-  .container .form-card .form-control h2 {
-    font-size: 2rem;
-  }
-  .container .form-card .form-control form {
-    display: flex;
-    flex-direction: column;
-    margin: 0px 30px;
-  }
-  .container .form-card .form-control form input {
-    margin: 10px 0px;
-    border: none;
-    padding: 15px;
-    background-color: #efefef;
-    border-radius: 5px;
-  }
- 
-  .container .form-card .form-control form button {
-    border: none;
-    padding: 20px;
-    margin-top: 5px;
-    background-color: #3f54be;
-    border-radius: 5px;
-    color: #fff;
-    cursor: pointer;
-  }
-  .container .form-card .form-control form button:focus {
-    outline: none;
-  }
-  .container .form-card .form-control span {
-    margin: 10px 0px;
-  }
-
-  .container .form-card .form-control .socials i {
-    margin: 0 5px;
-    color: #fff;
-    border-radius: 50%;
-  }
-  .container .form-card .form-control .socials .fa-facebook-f {
-    padding: 5px 8px;
-    background-color: #3b5998;
-  }
-  .container .form-card .form-control .socials .fa-google-plus-g {
-    padding: 5px 4px;
-    background-color: #db4a39;
-  }
-  .container .form-card .form-control .socials .fa-linkedin-in {
-    padding: 5px 6px;
-    background-color: #0e76a8;
-  }
-  .container .form-card .form-control.signup {
-    opacity: 0;
-    z-index: 1;
-    left: 200%;
-  }
-  .container .form-card .form-control.signin {
-    opacity: 1;
-    z-index: 2;
-    left: 0%;
-  }
-
-  .container .intros-container {
-    position: relative;
-    left: 50%;
-    width: 50%;
-    text-align: center;
-  }
-  .container .intros-container .intro-control {
-    position: absolute;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    height: 480px;
-    color: #fff;
-    background: linear-gradient(170deg, #34d399, #059669);
-    transition: all 0.5s ease-in;
-  }
-  .container .intros-container .intro-control .intro-control__inner {
-    margin: 0px 30px;
-  }
-  .container .intros-container .intro-control button {
-    border: none;
-    padding: 15px 30px;
-    background-color: #10b981;
-    border-radius: 50px;
-    color: #fff;
-    margin: 10px 0px;
-    cursor: pointer;
-  }
-  .container .intros-container .intro-control button:focus, .container .intros-container .intro-control button:hover {
-    outline: none;
-    background-color: #059669;
-  }
-
-.container .intros-container .intro-control h3,
-.container .intros-container .intro-control p {
-  margin: 10px 0px;
-}
-.container .intros-container .intro-control.signin-intro {
-  opacity: 1;
-  z-index: 2;
-}
-.container .intros-container .intro-control.signup-intro {
-  opacity: 0;
-  z-index: 1;
-}
-
-.change .form-card .form-control.signup {
-  opacity: 1;
-  z-index: 2;
-  transform: translateX(-100%);
-}
-.change .form-card .form-control.signup button {
-  background-color: #2563eb !important;
-}
-.change .form-card .form-control.signin {
-  opacity: 0;
-  z-index: 1;
-  transform: translateX(-100%);
-}
-
-.change .intros-container .intro-control {
-    transform: translateX(-100%);
-    background: linear-gradient(170deg, #3b82f6, #2563eb);
-  }
-  .change .intros-container .intro-control #signin-btn {
-    background-color: #2563eb;
-  }
-  .change .intros-container .intro-control.signin-intro {
-    opacity: 0;
-    z-index: 1;
-  }
-  .change .intros-container .intro-control.signup-intro {
-    opacity: 1;
-    z-index: 2;
-  }
-  
-  @media screen and (max-width: 480px) {
-    .container {
-      height: 100vh;
-      display: flex;
-      flex-direction: column;
-    }
-    .container .form-card {
-      order: 2;
-      width: 100%;
-      height: 50vh;
-    }
-    .container .form-card .form-control {
-      position: absolute;
-      height: 50vh;
-    }
-    .container .form-card .form-control.signup-form {
-      left: 0%;
-      margin-top: 70px;
-    }
-    .container .intros-container {
-      order: 1;
-      width: 100%;
-      left: 0%;
-      height: 40vh;
-    }
-    .container .intros-container .intro-control {
-      position: absolute;
-      height: 40vh;
-    }
-  
-    .change .form-card .form-control.signup {
-      transform: translateX(0%);
-    }
-    .change .form-card .form-control.signin {
-      transform: translateX(0%);
-    }
-    .change .intros-container .intro-control {
-      transform: translateX(0%);
-    }
-  }
